fixing a bunch of 64-bit size_t related compiler warnings (but nothing that looked like a major problem)
git-svn-id: https://gnunet.org/svn/GNUnet@1568 140774ce-b5e7-0310-ab8b-a85725594a96

+++ src/server/tcpserver.c
@@ -585,7 +585,7 @@ static void * tcpListenMain() {
 	}
       }
       if (FD_ISSET(sock, &writeSet)) {
-	int ret;
+	size_t ret;
 	
 #if DEBUG_TCPHANDLER
 	LOG(LOG_DEBUG,
@@ -635,7 +635,7 @@ try_again:
 	  pos = ch;
 	  continue;
 	}
-	if ((unsigned int)ret == pos->writeBufferSize) {
+	if (ret == pos->writeBufferSize) {
 	  FREENONNULL(pos->writeBuffer);
 	  pos->writeBuffer = NULL;
 	  pos->writeBufferSize = 0;
